reference cleanup
