+++ models/experimental/lingbot-va/test_inference.py
@@ -1,7 +1,30 @@
 import torch
 from wan_va.modules.model import WanTransformer3DModel
 from transformers import T5Tokenizer, T5EncoderModel
-from wan_va.utils import get_mesh_id  # Add this import
+
+
+def get_mesh_id(f, h, w, t, f_w=1, f_shift=0, action=False):
+    f_idx = torch.arange(f_shift, f + f_shift) * f_w
+    h_idx = torch.arange(h)
+    w_idx = torch.arange(w)
+    ff, hh, ww = torch.meshgrid(f_idx, h_idx, w_idx, indexing="ij")
+    if action:
+        ff_offset = (torch.ones([h]).cumsum(0) / (h + 1)).view(1, -1, 1)
+        ff = ff + ff_offset
+        hh = torch.ones_like(hh) * -1
+        ww = torch.ones_like(ww) * -1
+
+    grid_id = torch.cat(
+        [
+            ff.unsqueeze(0),
+            hh.unsqueeze(0),
+            ww.unsqueeze(0),
+        ],
+        dim=0,
+    ).flatten(1)
+    grid_id = torch.cat([grid_id, torch.full_like(grid_id[:1], t)], dim=0)
+    return grid_id
+
 
 # ----------------------------------------------------------------------
 # 1) Device / dtype
