--- models/experimental/lingbot-va/wan_va/__init__.py
@@ -1,2 +1,0 @@
-# Copyright 2024-2025 The Alibaba Wan Team Authors. All rights reserved.
-from . import distributed, modules
