--- models/experimental/lingbot-va/wan_va/distributed/__init__.py
@@ -1,1 +1,0 @@
-# Copyright 2024-2025 The Alibaba Wan Team Authors. All rights reserved.
