--- models/experimental/lingbot-va/wan_va/modules/__init__.py
@@ -1,4 +1,0 @@
-# Copyright 2024-2025 The Robbyant Team Authors. All rights reserved.
-from .utils import load_text_encoder, load_tokenizer, load_transformer, load_vae
-
-__all__ = ["load_transformer", "load_text_encoder", "load_tokenizer", "load_vae", "WanVAEStreamingWrapper"]
