--- models/experimental/lingbot-va/wan_va/utils/__init__.py
@@ -1,3 +1,0 @@
-from .utils import get_mesh_id
-
-__all__ = ["get_mesh_id"]
